Big Update
Designed ToyDataGenerator, which generates fake frames and correct answers. Built and trained simple Keras model on the toy data. Begun implementing final agent, which can read the its current state and current it to a frame. Need to implement using this frame and the loaded Keras model to take an action.